Code orgnization update

+++ src/Mathematics/NumberTheorem/GCD.cs
@@ -1,18 +1,20 @@
+using System.Numerics;
+
 namespace MyLibrary.Mathematics.NumberTheorem
 {
     public class GCD
     {
-        public int Value { get; set; }
-        public GCD(int a, int b)
+        public BigInteger Value { get; set; }
+        public GCD(BigInteger a, BigInteger b)
         {
             Value = EuclideanAlgorithm(a, b);
         }
-        public static int EuclideanAlgorithm(int a, int b)
+        public static BigInteger EuclideanAlgorithm(BigInteger a, BigInteger b)
         {
             if (b == 0) return a;
             return EuclideanAlgorithm(b, a % b);
         }
-        public static(int, int, int) ExtendedEuclideanAlgorithm(int a, int b)
+        public static(BigInteger, BigInteger, BigInteger) ExtendedEuclideanAlgorithm(BigInteger a, BigInteger b)
         {
             if (b == 0) return (1, 0, a);
             var(x, y, gcd) = ExtendedEuclideanAlgorithm(b, a % b);
